@@ -3,31 +3,33 @@
   "version": "5.4.1",
   "private": true,
   "dependencies": {
-    "@fluentui/react-components": "^9.19.1",
-    "@microsoft/teams-js": "^1.4.2",
+    "@fluentui/react": "^8.110.3",
+    "@fluentui/react-components": "^9.20.2",
+    "@fluentui/react-datepicker-compat": "^0.1.4",
+    "@fluentui/react-utilities": "^9.9.0",
+    "@microsoft/teams-js": "^2.11.0",
     "@reduxjs/toolkit": "^1.9.5",
     "@types/react-redux": "^7.1.25",
-    "@types/react-router-dom": "^5.1.8",
+    "@types/react-router-dom": "^5.3.3",
     "@types/validator": "^13.7.17",
     "adaptivecards": "^1.2.0",
-    "axios": "^0.21.4",
-    "color-hash": "^1.0.3",
-    "faker": "^4.1.0",
+    "color-hash": "^2.0.2",
+    "faker": "^5.5.3",
     "font-awesome": "^4.7.0",
-    "i18next": "^19.5.1",
-    "i18next-browser-languagedetector": "^5.0.0",
-    "i18next-http-backend": "^1.3.2",
+    "i18next": "^22.5.0",
+    "i18next-browser-languagedetector": "^7.0.1",
+    "i18next-http-backend": "^2.2.1",
     "markdown-it": "^13.0.1",
     "moment": "2.29.4",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-i18next": "^11.7.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-i18next": "^12.3.1",
     "react-redux": "^8.0.5",
-    "react-router-dom": "^5.2.0",
-    "sass": "^1.32.7",
-    "scheduler": "^0.20.0",
-    "typescript": "^3.8.2",
-    "typestyle": "^2.0.2",
+    "react-router-dom": "^6.11.2",
+    "sass": "^1.62.1",
+    "scheduler": "^0.20.2",
+    "typescript": "^3.9.7",
+    "typestyle": "^2.4.0",
     "validator": "^13.9.0"
   },
   "scripts": {
@@ -52,17 +54,21 @@
     ]
   },
   "devDependencies": {
-    "@fortawesome/fontawesome-free": "^5.8.2",
-    "@types/color-hash": "^1.0.0",
-    "@types/faker": "^4.1.5",
-    "@types/history": "^4.7.2",
-    "@types/jest": "24.0.13",
+    "@fortawesome/fontawesome-free": "^6.4.0",
+    "@types/color-hash": "^1.0.2",
+    "@types/history": "^4.7.11",
+    "@types/jest": "29.5.1",
     "@types/markdown-it": "^12.2.3",
-    "@types/node": "^12.0.4",
-    "@types/react": "^16.8.19",
-    "@types/react-dom": "^16.8.4",
-    "react-scripts": "^5.0.1",
-    "tslint": "^5.20.0",
-    "tslint-microsoft-contrib": "^6.2.0"
+    "@types/node": "^20.2.3",
+    "@types/react": "^18.2.7",
+    "@types/react-dom": "^18.2.4",
+    "@typescript-eslint/eslint-plugin": "^5.59.7",
+    "eslint": "^8.41.0",
+    "eslint-config-standard-with-typescript": "^34.0.1",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-n": "^15.7.0",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-react": "^7.32.2",
+    "react-scripts": "^5.0.1"
   }
 }
