@@ -1,12 +1,12 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import * as React from "react";
-import { useTranslation } from "react-i18next";
-import { Spinner } from "@fluentui/react-components";
-import { GetDraftMessagesAction } from "../../actions";
-import { RootState, useAppDispatch, useAppSelector } from "../../store";
-import { DraftMessageDetail } from "../MessageDetail/draftMessageDetail";
+import * as React from 'react';
+import { useTranslation } from 'react-i18next';
+import { Spinner } from '@fluentui/react-components';
+import { GetDraftMessagesAction } from '../../actions';
+import { RootState, useAppDispatch, useAppSelector } from '../../store';
+import { DraftMessageDetail } from './draftMessageDetail';
 
 export const DraftMessages = () => {
   const { t } = useTranslation();
@@ -18,12 +18,12 @@ export const DraftMessages = () => {
     if (draftMessages && draftMessages.length === 0) {
       GetDraftMessagesAction(dispatch);
     }
-  }, []);
+  }, [draftMessages]);
 
   return (
     <>
-      {loader && <Spinner labelPosition="below" label="Fetching..." />}
-      {draftMessages && draftMessages.length === 0 && !loader && <div>{t("EmptyDraftMessages")}</div>}
+      {loader && <Spinner labelPosition='below' label={t('fetching')} />}
+      {draftMessages && draftMessages.length === 0 && !loader && <div>{t('EmptyDraftMessages')}</div>}
       {draftMessages && draftMessages.length > 0 && !loader && <DraftMessageDetail draftMessages={draftMessages} />}
     </>
   );
