@@ -1,6 +1,8 @@
+/* eslint-disable @typescript-eslint/dot-notation */
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
+import './newMessage.scss';
 import * as AdaptiveCards from 'adaptivecards';
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
@@ -26,12 +28,14 @@ import {
   Textarea,
   tokens,
   useId,
+  Checkbox,
 } from '@fluentui/react-components';
 import { InfoLabel } from '@fluentui/react-components/unstable';
+import { TimePicker, DatePicker, IComboBox } from '@fluentui/react';
+import { initializeIcons } from '@fluentui/react/lib/Icons';
 import { ArrowUpload24Regular, Dismiss12Regular } from '@fluentui/react-icons';
-import * as microsoftTeams from '@microsoft/teams-js';
-
-import { GetDraftMessagesSilentAction, GetGroupsAction, GetTeamsDataAction, SearchGroupsAction, VerifyGroupAccessAction } from '../../actions';
+import { dialog } from '@microsoft/teams-js';
+import { GetDraftMessagesSilentAction, GetGroupsAction, GetTeamsDataAction, SearchGroupsAction, VerifyGroupAccessAction, GetScheduledMessagesSilentAction } from '../../actions';
 import { createDraftNotification, getDraftNotification, updateDraftNotification } from '../../apis/messageListApi';
 import { getBaseUrl } from '../../configVariables';
 import { RootState, useAppDispatch, useAppSelector } from '../../store';
@@ -51,6 +55,8 @@ interface IMessageState {
   rosters: any[];
   groups: any[];
   allUsers: boolean;
+  isScheduled?: boolean;
+  scheduledDate?: string;
 }
 
 interface ITeamTemplate {
@@ -101,8 +107,10 @@ let card: any;
 
 const MAX_SELECTED_TEAMS_NUM: number = 20;
 
+initializeIcons(/* optional base url */);
+
 export const NewMessage = () => {
-  let fileInput = React.createRef<any>();
+  const fileInput = React.createRef<any>();
   const { t } = useTranslation();
   const { id } = useParams() as any;
   const dispatch = useAppDispatch();
@@ -134,45 +142,37 @@ export const NewMessage = () => {
   const [teamsSelectedOptions, setTeamsSelectedOptions] = React.useState<ITeamTemplate[]>([]);
   const [rostersSelectedOptions, setRostersSelectedOptions] = React.useState<ITeamTemplate[]>([]);
   const [searchSelectedOptions, setSearchSelectedOptions] = React.useState<ITeamTemplate[]>([]);
+  const [scheduleSendCheckBox, setScheduleSendCheckBox] = React.useState(false);
+  const [scheduledDatePicker, setScheduledDatePicker] = React.useState(new Date());
+  const [scheduledTimePicker, setScheduledTimePicker] = React.useState(new Date());
 
   React.useEffect(() => {
     GetTeamsDataAction(dispatch);
     VerifyGroupAccessAction(dispatch);
   }, []);
 
-  React.useEffect(() => {
-    if (t) {
-      card = getInitAdaptiveCard(t('TitleText'));
-      setDefaultCard(card);
-      updateAdaptiveCard();
-    }
-  }, [t]);
-
   React.useEffect(
     () => {
-      setCardTitle(card, messageState.title);
-      setCardImageLink(card, messageState.imageLink);
-      setCardSummary(card, messageState.summary);
-      setCardAuthor(card, messageState.author);
-      setCardBtn(card, messageState.buttonTitle, messageState.buttonLink);
       if (!messageState.title && !messageState.imageLink && !messageState.summary && !messageState.author && !messageState.buttonTitle && !messageState.buttonLink) {
         card = getInitAdaptiveCard(t('TitleText') ?? '');
         setDefaultCard(card);
+      } else {
+        setCardTitle(card, messageState.title);
+        setCardImageLink(card, messageState.imageLink);
+        setCardSummary(card, messageState.summary);
+        setCardAuthor(card, messageState.author);
+        setCardBtn(card, messageState.buttonTitle, messageState.buttonLink);
       }
       updateAdaptiveCard();
     },
-    [messageState]);
+    [t, pageSelection, messageState]);
 
   React.useEffect(() => {
     if (id) {
       GetGroupsAction(dispatch, { id });
-      getDraftNotificationItem(id);
+      void getDraftNotificationItem(id);
     }
   }, [id]);
-
-  React.useEffect(() => {
-    updateAdaptiveCard();
-  }, [pageSelection]);
 
   React.useEffect(() => {
     setTeamsSelectedOptions([]);
@@ -197,7 +197,7 @@ export const NewMessage = () => {
   const getDraftNotificationItem = async (id: number) => {
     try {
       await getDraftNotification(id).then((response) => {
-        const draftMessageDetail = response.data;
+        const draftMessageDetail = response;
 
         if (draftMessageDetail.teams.length > 0) {
           setSelectedRadioButton(AudienceSelection.Teams);
@@ -221,15 +221,12 @@ export const NewMessage = () => {
           rosters: draftMessageDetail.rosters,
           groups: draftMessageDetail.groups,
           allUsers: draftMessageDetail.allUsers,
+          isScheduled: draftMessageDetail.isScheduled,
+          scheduledDate: draftMessageDetail.scheduledDate,
         });
-
-        setCardTitle(card, draftMessageDetail.title);
-        setCardImageLink(card, draftMessageDetail.imageLink);
-        setCardSummary(card, draftMessageDetail.summary);
-        setCardAuthor(card, draftMessageDetail.author);
-        setCardBtn(card, draftMessageDetail.buttonTitle, draftMessageDetail.buttonLink);
-
-        updateAdaptiveCard();
+        setScheduleSendCheckBox(draftMessageDetail.isScheduled);
+        setScheduledDatePicker(new Date(draftMessageDetail.scheduledDate));
+        setScheduledTimePicker(new Date(draftMessageDetail.scheduledDate));
       });
     } catch (error) {
       return error;
@@ -242,15 +239,38 @@ export const NewMessage = () => {
     const authorAsString = t('Author1');
     const buttonTitleAsString = t('ButtonTitle');
     setCardTitle(card, titleAsString);
-    let imgUrl = getBaseUrl() + '/image/imagePlaceholder.png';
+    const imgUrl = getBaseUrl() + '/image/imagePlaceholder.png';
     setCardImageLink(card, imgUrl);
     setCardSummary(card, summaryAsString);
     setCardAuthor(card, authorAsString);
     setCardBtn(card, buttonTitleAsString, 'https://adaptivecards.io');
   };
 
+  const handleScheduleSendCheckBox = (event: any) => {
+    setScheduleSendCheckBox((scheduleSendCheckBox) => !scheduleSendCheckBox);
+    if (event.target.value === 'on') {
+      setMessageState({ ...messageState, isScheduled: true });
+    } else {
+      setMessageState({ ...messageState, isScheduled: false });
+    }
+  };
+
+  // update the state variable whenever the date is changed in the date picker control
+  const handleScheduleSendDate = (selectedDate: Date | null | undefined) => {
+    if (selectedDate) {
+      setScheduledDatePicker(selectedDate);
+    }
+  };
+  // update the state variable whenever the time is changed in the time picker control
+  const handleScheduleSendTime = (_ev: React.FormEvent<IComboBox>, selectedTime: Date) => {
+    if (selectedTime) {
+      setScheduledTimePicker(selectedTime);
+      setMessageState({ ...messageState, scheduledDate: selectedTime.toISOString() });
+    }
+  };
+
   const updateAdaptiveCard = () => {
-    var adaptiveCard = new AdaptiveCards.AdaptiveCard();
+    const adaptiveCard = new AdaptiveCards.AdaptiveCard();
     adaptiveCard.parse(card);
     const renderCard = adaptiveCard.render();
     if (renderCard && pageSelection === CurrentPageSelection.CardCreation) {
@@ -274,9 +294,9 @@ export const NewMessage = () => {
   };
 
   const checkValidSizeOfImage = (resizedImageAsBase64: string) => {
-    var stringLength = resizedImageAsBase64.length - 'data:image/png;base64,'.length;
-    var sizeInBytes = 4 * Math.ceil(stringLength / 3) * 0.5624896334383812;
-    var sizeInKb = sizeInBytes / 1000;
+    const stringLength = resizedImageAsBase64.length - 'data:image/png;base64,'.length;
+    const sizeInBytes = 4 * Math.ceil(stringLength / 3) * 0.5624896334383812;
+    const sizeInKb = sizeInBytes / 1000;
 
     if (sizeInKb <= 1024) return true;
     else return false;
@@ -290,7 +310,7 @@ export const NewMessage = () => {
       const { type: mimeType } = file;
 
       if (!validImageTypes.includes(fileType)) {
-        setImageUploadErrorMessage(t('ErrorImageTypesMessage'));
+        setImageUploadErrorMessage(t('ErrorImageTypesMessage') ?? '');
         return;
       }
 
@@ -300,9 +320,9 @@ export const NewMessage = () => {
       const fileReader = new FileReader();
       fileReader.readAsDataURL(file);
       fileReader.onload = () => {
-        var image = new Image();
+        const image = new Image();
         image.src = fileReader.result as string;
-        var resizedImageAsBase64 = fileReader.result as string;
+        let resizedImageAsBase64 = fileReader.result as string;
 
         image.onload = function (e: any) {
           const MAX_WIDTH = 1024;
@@ -321,27 +341,24 @@ export const NewMessage = () => {
         };
 
         if (!checkValidSizeOfImage(resizedImageAsBase64)) {
-          setImageUploadErrorMessage(t('ErrorImageSizeMessage'));
+          setImageUploadErrorMessage(t('ErrorImageSizeMessage') ?? '');
           return;
         }
 
-        setCardImageLink(card, resizedImageAsBase64);
         setMessageState({ ...messageState, imageLink: resizedImageAsBase64 });
-
-        updateAdaptiveCard();
       };
     }
   };
 
   const isSaveBtnDisabled = () => {
-    const msg_page_conditions = messageState.title !== '' && imageUploadErrorMessage === '' && btnLinkErrorMessage === '';
-    const aud_page_conditions =
+    const msgPageConditions = messageState.title !== '' && imageUploadErrorMessage === '' && btnLinkErrorMessage === '';
+    const audPageConditions =
       (teamsSelectedOptions.length > 0 && selectedRadioButton === AudienceSelection.Teams) ||
       (rostersSelectedOptions.length > 0 && selectedRadioButton === AudienceSelection.Rosters) ||
       (searchSelectedOptions.length > 0 && selectedRadioButton === AudienceSelection.Groups) ||
       selectedRadioButton === AudienceSelection.AllUsers;
 
-    if (msg_page_conditions && aud_page_conditions) {
+    if (msgPageConditions && audPageConditions) {
       return false;
     } else {
       return true;
@@ -396,11 +413,15 @@ export const NewMessage = () => {
     try {
       updateDraftNotification(msg)
         .then(() => {
-          GetDraftMessagesSilentAction(dispatch);
+          if (msg.isScheduled) {
+            GetScheduledMessagesSilentAction(dispatch);
+          } else {
+            GetDraftMessagesSilentAction(dispatch);
+          }
         })
         .finally(() => {
           setShowMsgDraftingSpinner(false);
-          microsoftTeams.tasks.submitTask();
+          dialog.url.submit();
         });
     } catch (error) {
       return error;
@@ -411,11 +432,15 @@ export const NewMessage = () => {
     try {
       createDraftNotification(msg)
         .then(() => {
-          GetDraftMessagesSilentAction(dispatch);
+          if (msg.isScheduled) {
+            GetScheduledMessagesSilentAction(dispatch);
+          } else {
+            GetDraftMessagesSilentAction(dispatch);
+          }
         })
         .finally(() => {
           setShowMsgDraftingSpinner(false);
-          microsoftTeams.tasks.submitTask();
+          dialog.url.submit();
         });
     } catch (error) {
       return error;
@@ -426,6 +451,10 @@ export const NewMessage = () => {
     setPageSelection(CurrentPageSelection.AudienceSelection);
   };
 
+  const onCancel = (event: any) => {
+    dialog.url.submit();
+  };
+
   const onBack = (event: any) => {
     setPageSelection(CurrentPageSelection.CardCreation);
     setAllUserAria('none');
@@ -434,13 +463,11 @@ export const NewMessage = () => {
 
   const onTitleChanged = (event: any) => {
     if (event.target.value === '') {
-      setTitleErrorMessage('Title is required.');
+      setTitleErrorMessage(t('titleRequired') ?? '');
     } else {
       setTitleErrorMessage('');
     }
-    setCardTitle(card, event.target.value);
     setMessageState({ ...messageState, title: event.target.value });
-    updateAdaptiveCard();
   };
 
   const onImageLinkChanged = (event: any) => {
@@ -458,46 +485,37 @@ export const NewMessage = () => {
       )
     ) {
       isGoodLink = false;
-      setImageUploadErrorMessage(t('ErrorURLMessage'));
+      setImageUploadErrorMessage(t('ErrorURLMessage') ?? '');
     } else {
       isGoodLink = true;
-      setImageUploadErrorMessage(t(''));
+      setImageUploadErrorMessage('');
     }
 
     if (isGoodLink) {
       setMessageState({ ...messageState, imageLink: urlOrDataUrl });
-      setCardImageLink(card, event.target.value);
-      updateAdaptiveCard();
     }
   };
 
   const onSummaryChanged = (event: any) => {
-    setCardSummary(card, event.target.value);
     setMessageState({ ...messageState, summary: event.target.value });
-    updateAdaptiveCard();
   };
 
   const onAuthorChanged = (event: any) => {
-    setCardAuthor(card, event.target.value);
     setMessageState({ ...messageState, author: event.target.value });
-    updateAdaptiveCard();
   };
 
   const onBtnTitleChanged = (event: any) => {
-    setCardBtn(card, event.target.value, messageState.buttonLink);
     setMessageState({ ...messageState, buttonTitle: event.target.value });
-    updateAdaptiveCard();
   };
 
   const onBtnLinkChanged = (event: any) => {
     if (validator.isURL(event.target.value, { require_protocol: true, protocols: ['https'] }) || event.target.value === '') {
       setBtnLinkErrorMessage('');
     } else {
-      setBtnLinkErrorMessage(`${event.target.value} is invalid. Please enter a valid https URL`);
+      // eslint-disable-next-line @typescript-eslint/restrict-template-expressions
+      setBtnLinkErrorMessage(`${event.target.value} is invalid. ${t('enterValidURL')}`);
     }
-    setCardBtn(card, messageState.buttonTitle, event.target.value);
     setMessageState({ ...messageState, buttonLink: event.target.value });
-    updateAdaptiveCard();
   };
 
   // generate ids for handling labelling
@@ -539,7 +557,7 @@ export const NewMessage = () => {
   };
 
   const onSearchChange = (event: any) => {
-    if (event && event.target && event.target.value) {
+    if (event?.target?.value) {
       const q = encodeURIComponent(event.target.value);
       SearchGroupsAction(dispatch, { query: q });
     }
@@ -589,14 +607,13 @@ export const NewMessage = () => {
 
   const teamsLabelledBy = teamsSelectedOptions.length > 0 ? `${teamsComboId} ${teamsSelectedListId}` : teamsComboId;
   const rostersLabelledBy = rostersSelectedOptions.length > 0 ? `${rostersComboId} ${rostersSelectedListId}` : rostersComboId;
-
   const searchLabelledBy = searchSelectedOptions.length > 0 ? `${searchComboId} ${searchSelectedListId}` : searchComboId;
 
-  const cmb_styles = useComboboxStyles();
-  const field_styles = useFieldStyles();
+  const cmbStyles = useComboboxStyles();
+  const fieldStyles = useFieldStyles();
 
   const audienceSelectionChange = (ev: any, data: RadioGroupOnChangeData) => {
-    let input = data.value as keyof typeof AudienceSelection;
+    const input = data.value as keyof typeof AudienceSelection;
     setSelectedRadioButton(AudienceSelection[input]);
 
     if (AudienceSelection[input] === AudienceSelection.AllUsers) {
@@ -613,12 +630,13 @@ export const NewMessage = () => {
     <>
       {pageSelection === CurrentPageSelection.CardCreation && (
         <>
-          <span role='alert' aria-label={t('NewMessageStep1')} />
+          <span role='alert' aria-label={t('NewMessageStep1') ?? ''} />
           <div className='adaptive-task-grid'>
             <div className='form-area'>
-              <Field size='large' className={field_styles.styles} label={t('TitleText')} required={true} validationMessage={titleErrorMessage}>
+              <Field size='large' className={fieldStyles.styles} label={t('TitleText')} required={true} validationMessage={titleErrorMessage}>
                 <Input
-                  placeholder={t('PlaceHolderTitle')}
+                  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                  placeholder={t('PlaceHolderTitle')!}
                   onChange={onTitleChanged}
                   autoComplete='off'
                   size='large'
@@ -629,10 +647,12 @@ export const NewMessage = () => {
               </Field>
               <Field
                 size='large'
-                className={field_styles.styles}
+                className={fieldStyles.styles}
+                // eslint-disable-next-line @typescript-eslint/ban-ts-comment, @typescript-eslint/prefer-ts-expect-error
+                // @ts-ignore
                 label={{
                   children: (_: unknown, imageInfoProps: LabelProps) => (
-                    <InfoLabel {...imageInfoProps} info={t('ImageSizeInfoContent') || ''}>
+                    <InfoLabel {...imageInfoProps} info={t('ImageSizeInfoContent') ?? ''}>
                       {t('ImageURL')}
                     </InfoLabel>
                   ),
@@ -651,22 +671,40 @@ export const NewMessage = () => {
                     style={{ gridColumn: '1' }}
                     appearance='filled-darker'
                     value={imageFileName || ''}
-                    placeholder={t('ImageURL')}
+                    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                    placeholder={t('PlaceHolderImageURL')!}
                     onChange={onImageLinkChanged}
                   />
-                  <Button
-                    style={{ gridColumn: '2', marginLeft: '5px' }}
-                    onClick={handleUploadClick}
-                    size='large'
-                    appearance='secondary'
-                    aria-label={imageFileName ? t('UploadImageSuccessful') : t('UploadImageInfo')}
-                    icon={<ArrowUpload24Regular />}
-                  >
-                    {t('Upload')}
-                  </Button>
+                  <div
+                    style={{
+                      gridColumn: '2',
+                      marginLeft: '8px',
+                      marginRight: '5px',
+                      paddingTop: '8px',
+                      color: 'darkgray',
+                    }}
+                    placeholder="{t('FieldSeperator')}"
+                  >{' '}(Or){' '}
+                  </div>
+                  {
+                    // eslint-disable-next-line @typescript-eslint/ban-ts-comment, @typescript-eslint/prefer-ts-expect-error
+                    // @ts-ignore
+                    <Button
+                      style={{ gridColumn: '3', marginLeft: '5px' }}
+                      onClick={handleUploadClick}
+                      size='large'
+                      appearance='secondary'
+                      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                      aria-label={imageFileName ? t('UploadImageSuccessful') : t('UploadImageInfo')}
+                      icon={<ArrowUpload24Regular />}
+                    >
+                      {t('Upload')}
+                    </Button>
+                  }
                   <input
                     type='file'
                     accept='.jpg, .jpeg, .png, .gif'
+                    aria-label='input file upload (hidden)'
                     style={{ display: 'none' }}
                     multiple={false}
                     onChange={handleImageSelection}
@@ -674,44 +712,49 @@ export const NewMessage = () => {
                   />
                 </div>
               </Field>
-              <Field size='large' className={field_styles.styles} label={t('Summary')}>
+              <Field size='large' className={fieldStyles.styles} label={t('Summary')}>
                 <Textarea
                   size='large'
+                  resize='vertical'
                   appearance='filled-darker'
-                  placeholder={t('Summary')}
-                  value={messageState.summary || ''}
+                  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                  placeholder={t('PlaceHolderSummary')!}
+                  value={messageState.summary ?? ''}
                   onChange={onSummaryChanged}
                 />
               </Field>
-              <Field size='large' className={field_styles.styles} label={t('Author')}>
+              <Field size='large' className={fieldStyles.styles} label={t('Author')}>
                 <Input
-                  placeholder={t('Author')}
+                  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                  placeholder={t('PlaceHolderAuthor')!}
                   size='large'
                   onChange={onAuthorChanged}
                   autoComplete='off'
                   appearance='filled-darker'
-                  value={messageState.author || ''}
+                  value={messageState.author ?? ''}
                 />
               </Field>
-              <Field size='large' className={field_styles.styles} label={t('ButtonTitle')}>
+              <Field size='large' className={fieldStyles.styles} label={t('ButtonTitle')}>
                 <Input
                   size='large'
-                  placeholder={t('ButtonTitle')}
+                  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                  placeholder={t('PlaceHolderButtonTitle')!}
                   onChange={onBtnTitleChanged}
                   autoComplete='off'
                   appearance='filled-darker'
-                  value={messageState.buttonTitle || ''}
+                  value={messageState.buttonTitle ?? ''}
                 />
               </Field>
-              <Field size='large' className={field_styles.styles} label={t('ButtonURL')} validationMessage={btnLinkErrorMessage}>
+              <Field size='large' className={fieldStyles.styles} label={t('ButtonURL')} validationMessage={btnLinkErrorMessage}>
                 <Input
                   size='large'
-                  placeholder={t('ButtonURL')}
+                  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                  placeholder={t('PlaceHolderButtonURL')!}
                   onChange={onBtnLinkChanged}
                   type='url'
                   autoComplete='off'
                   appearance='filled-darker'
-                  value={messageState.buttonLink || ''}
+                  value={messageState.buttonLink ?? ''}
                 />
               </Field>
             </div>
@@ -723,6 +766,9 @@ export const NewMessage = () => {
           </div>
           <div className='fixed-footer'>
             <div className='footer-action-right'>
+              <Button id='cancelBtn' onClick={onCancel} appearance='secondary' style={{ marginRight: '16px' }}>
+                {t('Cancel')}
+              </Button>
               <Button disabled={isNextBtnDisabled()} id='saveBtn' onClick={onNext} appearance='primary'>
                 {t('Next')}
               </Button>
@@ -732,8 +778,8 @@ export const NewMessage = () => {
       )}
       {pageSelection === CurrentPageSelection.AudienceSelection && (
         <>
-          <span role='alert' aria-label={t('NewMessageStep2')} />
-          <div className='adaptive-task-grid'>
+          <span role='alert' aria-label={t('NewMessageStep2') ?? ''} />
+          <div className='adaptive-task-grid new-messages'>
             <div className='form-area'>
               <Label size='large' id='audienceSelectionGroupLabelId'>
                 {t('SendHeadingText')}
@@ -741,34 +787,38 @@ export const NewMessage = () => {
               <RadioGroup defaultValue={selectedRadioButton} aria-labelledby='audienceSelectionGroupLabelId' onChange={audienceSelectionChange}>
                 <Radio id='radio1' value={AudienceSelection.Teams} label={t('SendToGeneralChannel')} />
                 {selectedRadioButton === AudienceSelection.Teams && (
-                  <div className={cmb_styles.root}>
-                    <Label id={teamsComboId}>Pick team(s)</Label>
-                    {teamsSelectedOptions.length ? (
-                      <ul id={teamsSelectedListId} className={cmb_styles.tagsList} ref={teamsSelectedListRef}>
-                        {/* The "Remove" span is used for naming the buttons without affecting the Combobox name */}
-                        <span id={`${teamsComboId}-remove`} hidden>
-                          Remove
-                        </span>
-                        {teamsSelectedOptions.map((option, i) => (
-                          <li key={option.id}>
-                            <Button
-                              size='small'
-                              shape='rounded'
-                              appearance='subtle'
-                              icon={<Dismiss12Regular />}
-                              iconPosition='after'
-                              onClick={() => onTeamsTagClick(option, i)}
-                              id={`${teamsComboId}-remove-${i}`}
-                              aria-labelledby={`${teamsComboId}-remove ${teamsComboId}-remove-${i}`}
-                            >
-                              <Persona name={option.name} secondaryText={'Team'} avatar={{ shape: 'square', color: 'colorful' }} />
-                            </Button>
-                          </li>
-                        ))}
-                      </ul>
-                    ) : (
-                      <></>
-                    )}
+                  <div className={cmbStyles.root}>
+                    <Label id={teamsComboId}>{t('pickTeams')}</Label>
+                    {
+                      // eslint-disable-next-line multiline-ternary
+                      teamsSelectedOptions.length ? (
+                        <ul id={teamsSelectedListId} className={cmbStyles.tagsList} ref={teamsSelectedListRef}>
+                          {/* The "Remove" span is used for naming the buttons without affecting the Combobox name */}
+                          <span id={`${teamsComboId}-remove`} hidden>
+                            {t('remove')}
+                          </span>
+                          {teamsSelectedOptions.map((option, i) => (
+                            <li key={option.id}>
+                              <Button
+                                size='small'
+                                shape='rounded'
+                                appearance='subtle'
+                                icon={<Dismiss12Regular />}
+                                iconPosition='after'
+                                // eslint-disable-next-line @typescript-eslint/no-confusing-void-expression
+                                onClick={() => onTeamsTagClick(option, i)}
+                                id={`${teamsComboId}-remove-${i}`}
+                                aria-labelledby={`${teamsComboId}-remove ${teamsComboId}-remove-${i}`}
+                              >
+                                <Persona name={option.name} secondaryText={'Team'} avatar={{ shape: 'square', color: 'colorful' }} />
+                              </Button>
+                            </li>
+                          ))}
+                        </ul>
+                      ) : (
+                        <></>
+                      )
+                    }
                     <Combobox
                       multiselect={true}
                       selectedOptions={teamsSelectedOptions.map((op) => op.id)}
@@ -777,7 +827,8 @@ export const NewMessage = () => {
                       onOptionSelect={onTeamsSelect}
                       ref={teamsComboboxInputRef}
                       aria-labelledby={teamsLabelledBy}
-                      placeholder={teams.length !== 0 ? 'Pick one or more teams' : t('NoMatchMessage')}
+                      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                      placeholder={teams.length !== 0 ? t('pickOneOrMoreTeams')! : t('NoMatchMessage')!}
                     >
                       {teams.map((opt) => (
                         <Option text={opt.name} value={opt.id} key={opt.id}>
@@ -789,34 +840,38 @@ export const NewMessage = () => {
                 )}
                 <Radio id='radio2' value={AudienceSelection.Rosters} label={t('SendToRosters')} />
                 {selectedRadioButton === AudienceSelection.Rosters && (
-                  <div className={cmb_styles.root}>
-                    <Label id={rostersComboId}>Pick team(s)</Label>
-                    {rostersSelectedOptions.length ? (
-                      <ul id={rostersSelectedListId} className={cmb_styles.tagsList} ref={rostersSelectedListRef}>
-                        {/* The "Remove" span is used for naming the buttons without affecting the Combobox name */}
-                        <span id={`${rostersComboId}-remove`} hidden>
-                          Remove
-                        </span>
-                        {rostersSelectedOptions.map((option, i) => (
-                          <li key={option.id}>
-                            <Button
-                              size='small'
-                              shape='rounded'
-                              appearance='subtle'
-                              icon={<Dismiss12Regular />}
-                              iconPosition='after'
-                              onClick={() => onRostersTagClick(option, i)}
-                              id={`${rostersComboId}-remove-${i}`}
-                              aria-labelledby={`${rostersComboId}-remove ${rostersComboId}-remove-${i}`}
-                            >
-                              <Persona name={option.name} secondaryText={'Team'} avatar={{ shape: 'square', color: 'colorful' }} />
-                            </Button>
-                          </li>
-                        ))}
-                      </ul>
-                    ) : (
-                      <></>
-                    )}
+                  <div className={cmbStyles.root}>
+                    <Label id={rostersComboId}>{t('pickTeams')}</Label>
+                    {
+                      // eslint-disable-next-line multiline-ternary
+                      rostersSelectedOptions.length ? (
+                        <ul id={rostersSelectedListId} className={cmbStyles.tagsList} ref={rostersSelectedListRef}>
+                          {/* The "Remove" span is used for naming the buttons without affecting the Combobox name */}
+                          <span id={`${rostersComboId}-remove`} hidden>
+                            {t('remove')}
+                          </span>
+                          {rostersSelectedOptions.map((option, i) => (
+                            <li key={option.id}>
+                              <Button
+                                size='small'
+                                shape='rounded'
+                                appearance='subtle'
+                                icon={<Dismiss12Regular />}
+                                iconPosition='after'
+                                // eslint-disable-next-line @typescript-eslint/no-confusing-void-expression
+                                onClick={() => onRostersTagClick(option, i)}
+                                id={`${rostersComboId}-remove-${i}`}
+                                aria-labelledby={`${rostersComboId}-remove ${rostersComboId}-remove-${i}`}
+                              >
+                                <Persona name={option.name} secondaryText={'Team'} avatar={{ shape: 'square', color: 'colorful' }} />
+                              </Button>
+                            </li>
+                          ))}
+                        </ul>
+                      ) : (
+                        <></>
+                      )
+                    }
                     <Combobox
                       multiselect={true}
                       selectedOptions={rostersSelectedOptions.map((op) => op.id)}
@@ -825,7 +880,8 @@ export const NewMessage = () => {
                       onOptionSelect={onRostersSelect}
                       ref={rostersComboboxInputRef}
                       aria-labelledby={rostersLabelledBy}
-                      placeholder={teams.length !== 0 ? 'Pick one or more teams' : t('NoMatchMessage')}
+                      // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+                      placeholder={teams.length !== 0 ? t('pickOneOrMoreTeams')! : t('NoMatchMessage')!}
                     >
                       {teams.map((opt) => (
                         <Option text={opt.name} value={opt.id} key={opt.id}>
@@ -836,7 +892,7 @@ export const NewMessage = () => {
                   </div>
                 )}
                 <Radio id='radio3' value={AudienceSelection.AllUsers} label={t('SendToAllUsers')} />
-                <div className={cmb_styles.root}>
+                <div className={cmbStyles.root}>
                   {selectedRadioButton === AudienceSelection.AllUsers && (
                     <Text id='radio3Note' role={allUsersAria} className='info-text'>
                       {t('SendToAllUsersNote')}
@@ -845,7 +901,7 @@ export const NewMessage = () => {
                 </div>
                 <Radio id='radio4' value={AudienceSelection.Groups} label={t('SendToGroups')} />
                 {selectedRadioButton === AudienceSelection.Groups && (
-                  <div className={cmb_styles.root}>
+                  <div className={cmbStyles.root}>
                     {!canAccessGroups && (
                       <Text role={groupsAria} className='info-text'>
                         {t('SendToGroupsPermissionNote')}
@@ -853,40 +909,44 @@ export const NewMessage = () => {
                     )}
                     {canAccessGroups && (
                       <>
-                        <Label id={searchComboId}>Pick group(s)</Label>
-                        {searchSelectedOptions.length ? (
-                          <ul id={searchSelectedListId} className={cmb_styles.tagsList} ref={searchSelectedListRef}>
-                            {/* The "Remove" span is used for naming the buttons without affecting the Combobox name */}
-                            <span id={`${searchComboId}-remove`} hidden>
-                              Remove
-                            </span>
-                            {searchSelectedOptions.map((option, i) => (
-                              <li key={option.id}>
-                                <Button
-                                  size='small'
-                                  shape='rounded'
-                                  appearance='subtle'
-                                  icon={<Dismiss12Regular />}
-                                  iconPosition='after'
-                                  onClick={() => onSearchTagClick(option, i)}
-                                  id={`${searchComboId}-remove-${i}`}
-                                  aria-labelledby={`${searchComboId}-remove ${searchComboId}-remove-${i}`}
-                                >
-                                  <Persona name={option.name} secondaryText={'Group'} avatar={{ color: 'colorful' }} />
-                                </Button>
-                              </li>
-                            ))}
-                          </ul>
-                        ) : (
-                          <></>
-                        )}
+                        <Label id={searchComboId}>{t('pickGroups')}</Label>
+                        {
+                          // eslint-disable-next-line multiline-ternary
+                          searchSelectedOptions.length ? (
+                            <ul id={searchSelectedListId} className={cmbStyles.tagsList} ref={searchSelectedListRef}>
+                              {/* The "Remove" span is used for naming the buttons without affecting the Combobox name */}
+                              <span id={`${searchComboId}-remove`} hidden>
+                                {t('remove')}
+                              </span>
+                              {searchSelectedOptions.map((option, i) => (
+                                <li key={option.id}>
+                                  <Button
+                                    size='small'
+                                    shape='rounded'
+                                    appearance='subtle'
+                                    icon={<Dismiss12Regular />}
+                                    iconPosition='after'
+                                    // eslint-disable-next-line @typescript-eslint/no-confusing-void-expression
+                                    onClick={() => onSearchTagClick(option, i)}
+                                    id={`${searchComboId}-remove-${i}`}
+                                    aria-labelledby={`${searchComboId}-remove ${searchComboId}-remove-${i}`}
+                                  >
+                                    <Persona name={option.name} secondaryText={'Group'} avatar={{ color: 'colorful' }} />
+                                  </Button>
+                                </li>
+                              ))}
+                            </ul>
+                          ) : (
+                            <></>
+                          )
+                        }
                         <Combobox
                           appearance='filled-darker'
                           size='large'
                           onOptionSelect={onSearchSelect}
                           onChange={onSearchChange}
                           aria-labelledby={searchLabelledBy}
-                          placeholder={'Search for groups'}
+                          placeholder={t('searchForGroups') ?? ''}
                         >
                           {queryGroups.map((opt) => (
                             <Option text={opt.name} value={opt.id} key={opt.id}>
@@ -902,6 +962,60 @@ export const NewMessage = () => {
                   </div>
                 )}
               </RadioGroup>
+              <div>
+                <></>
+              </div>
+              <div>
+                <Label size='large' id='MoreOptionsLabelId'>
+                  {t('MoreOptions')}
+                </Label>
+              </div>
+              <Checkbox
+                id='ScheduleCheckbox'
+                label={t('ScheduleSend')}
+                defaultChecked={scheduleSendCheckBox}
+                onChange={handleScheduleSendCheckBox}
+              />
+              {scheduleSendCheckBox && (
+                <div>
+                  <Label
+                    size='small'
+                    id='ScheduleSection'
+                    className='info-text'
+                    style={{ marginBottom: '5px', display: 'block', marginLeft: '36px' }}
+                  >
+                    {t('ScheduleSection')}
+                  </Label>
+                  <Label
+                    size='small'
+                    id='ScheduleNote'
+                    className='info-text'
+                    style={{ marginBottom: '5px', display: 'block', marginLeft: '36px' }}
+                  >
+                    {t('ScheduleNote')}
+                  </Label>
+                  <div className='schedulesend-datetime'>
+                    <DatePicker
+                      value={scheduledDatePicker ?? new Date()}
+                      onSelectDate={handleScheduleSendDate}
+                      minDate={new Date()}
+                      placeholder='Select a date'
+                      ariaLabel={'Scheduled Date required'}
+                      className='schedule-datepicker'
+                      calloutProps={{ className: 'incidentdatepicker-callout' }}
+                    />
+                    <TimePicker
+                      dateAnchor={scheduledDatePicker}
+                      value={scheduledTimePicker ?? new Date()}
+                      placeholder='Select a time'
+                      onChange={handleScheduleSendTime}
+                      calloutProps={{ directionalHintFixed: true, doNotLayer: true }}
+                      ariaLabel={'Scheduled Time required'}
+                      className='schedule-timepicker'
+                    />
+                  </div>
+                </div>
+              )}
             </div>
             <div className='card-area'>
               <div className={cardAreaBorderClass}>
