@@ -6,13 +6,9 @@ import * as React from 'react';
 import { useTranslation } from 'react-i18next';
 import { useParams } from 'react-router-dom';
 import { Button, Field, Label, Persona, Spinner, Text } from '@fluentui/react-components';
-import * as microsoftTeams from '@microsoft/teams-js';
-import {
-    getConsentSummaries, getDraftNotification, sendDraftNotification
-} from '../../apis/messageListApi';
-import {
-    getInitAdaptiveCard, setCardAuthor, setCardBtn, setCardImageLink, setCardSummary, setCardTitle
-} from '../AdaptiveCard/adaptiveCard';
+import { dialog } from '@microsoft/teams-js';
+import { getConsentSummaries, getDraftNotification, sendDraftNotification } from '../../apis/messageListApi';
+import { getInitAdaptiveCard, setCardAuthor, setCardBtn, setCardImageLink, setCardSummary, setCardTitle } from '../AdaptiveCard/adaptiveCard';
 import { AvatarShape } from '@fluentui/react-avatar';
 
 export interface IMessageState {
@@ -54,8 +50,8 @@ export const SendConfirmationTask = () => {
   const [cardAreaBorderClass, setCardAreaBorderClass] = React.useState('');
 
   const [messageState, setMessageState] = React.useState<IMessageState>({
-    id: "",
-    title: "",
+    id: '',
+    title: '',
     isDraftMsgUpdated: false,
   });
 
@@ -70,22 +66,22 @@ export const SendConfirmationTask = () => {
 
   React.useEffect(() => {
     if (id) {
-      getDraftMessage(id);
-      getConsents(id);
+      void getDraftMessage(id);
+      void getConsents(id);
     }
   }, [id]);
 
   React.useEffect(() => {
     if (isCardReady && consentState.isConsentsUpdated && messageState.isDraftMsgUpdated) {
-      var adaptiveCard = new AdaptiveCards.AdaptiveCard();
+      const adaptiveCard = new AdaptiveCards.AdaptiveCard();
       adaptiveCard.parse(card);
       const renderCard = adaptiveCard.render();
       if (renderCard) {
-        document.getElementsByClassName("card-area-1")[0].appendChild(renderCard);
+        document.getElementsByClassName('card-area-1')[0].appendChild(renderCard);
         setCardAreaBorderClass('card-area-border');
       }
       adaptiveCard.onExecuteAction = function (action: any) {
-        window.open(action.url, "_blank");
+        window.open(action.url, '_blank');
       };
       setLoader(false);
     }
@@ -106,8 +102,8 @@ export const SendConfirmationTask = () => {
   const getDraftMessage = async (id: number) => {
     try {
       await getDraftNotification(id).then((response) => {
-        updateCardData(response.data);
-        setMessageState({ ...response.data, isDraftMsgUpdated: true });
+        updateCardData(response);
+        setMessageState({ ...response, isDraftMsgUpdated: true });
       });
     } catch (error) {
       return error;
@@ -119,10 +115,10 @@ export const SendConfirmationTask = () => {
       await getConsentSummaries(id).then((response) => {
         setConsentState({
           ...consentState,
-          teamNames: response.data.teamNames.sort(),
-          rosterNames: response.data.rosterNames.sort(),
-          groupNames: response.data.groupNames.sort(),
-          allUsers: response.data.allUsers,
+          teamNames: response.teamNames.sort(),
+          rosterNames: response.rosterNames.sort(),
+          groupNames: response.groupNames.sort(),
+          allUsers: response.allUsers,
           messageId: id,
           isConsentsUpdated: true,
         });
@@ -136,7 +132,7 @@ export const SendConfirmationTask = () => {
     setDisableSendButton(true);
     sendDraftNotification(messageState)
       .then(() => {
-        microsoftTeams.tasks.submitTask();
+        dialog.url.submit();
       })
       .finally(() => {
         setDisableSendButton(false);
@@ -144,12 +140,13 @@ export const SendConfirmationTask = () => {
   };
 
   const getItemList = (items: string[], secondaryText: string, shape: AvatarShape) => {
-    let resultedTeams: any[] = [];
+    const resultedTeams: any[] = [];
     if (items) {
+      // eslint-disable-next-line array-callback-return
       items.map((element) => {
         resultedTeams.push(
-          <li key={element + "key"}>
-            <Persona name={element} secondaryText={secondaryText} avatar={{ shape, color: "colorful" }} />
+          <li key={element + 'key'}>
+            <Persona name={element} secondaryText={secondaryText} avatar={{ shape, color: 'colorful' }} />
           </li>
         );
       });
@@ -160,31 +157,31 @@ export const SendConfirmationTask = () => {
   const renderAudienceSelection = () => {
     if (consentState.teamNames && consentState.teamNames.length > 0) {
       return (
-        <div key="teamNames" style={{ paddingBottom: "16px" }}>
-          <Label>{t("TeamsLabel")}</Label>
-          <ul className="ul-no-bullets">{getItemList(consentState.teamNames, "Team", "square")}</ul>
+        <div key='teamNames' style={{ paddingBottom: '16px' }}>
+          <Label>{t('TeamsLabel')}</Label>
+          <ul className='ul-no-bullets'>{getItemList(consentState.teamNames, 'Team', 'square')}</ul>
         </div>
       );
     } else if (consentState.rosterNames && consentState.rosterNames.length > 0) {
       return (
-        <div key="rosterNames" style={{ paddingBottom: "16px" }}>
-          <Label>{t("TeamsMembersLabel")}</Label>
-          <ul className="ul-no-bullets">{getItemList(consentState.rosterNames, "Team", "square")}</ul>
+        <div key='rosterNames' style={{ paddingBottom: '16px' }}>
+          <Label>{t('TeamsMembersLabel')}</Label>
+          <ul className='ul-no-bullets'>{getItemList(consentState.rosterNames, 'Team', 'square')}</ul>
         </div>
       );
     } else if (consentState.groupNames && consentState.groupNames.length > 0) {
       return (
-        <div key="groupNames" style={{ paddingBottom: "16px" }}>
-          <Label>{t("GroupsMembersLabel")}</Label>
-          <ul className="ul-no-bullets">{getItemList(consentState.groupNames, "Group", "circular")}</ul>
+        <div key='groupNames' style={{ paddingBottom: '16px' }}>
+          <Label>{t('GroupsMembersLabel')}</Label>
+          <ul className='ul-no-bullets'>{getItemList(consentState.groupNames, 'Group', 'circular')}</ul>
         </div>
       );
     } else if (consentState.allUsers) {
       return (
-        <div key="allUsers" style={{ paddingBottom: "16px" }}>
-          <Label>{t("AllUsersLabel")}</Label>
+        <div key='allUsers' style={{ paddingBottom: '16px' }}>
+          <Label>{t('AllUsersLabel')}</Label>
           <div>
-            <Text className="info-text">{t("SendToAllUsersNote")}</Text>
+            <Text className='info-text'>{t('SendToAllUsersNote')}</Text>
           </div>
         </div>
       );
