@@ -1,39 +1,40 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import React from "react";
-import * as microsoftTeams from "@microsoft/teams-js";
-import { getBaseUrl } from "../configVariables";
+import React from 'react';
+import { app, pages } from '@microsoft/teams-js';
+import { getBaseUrl } from '../configVariables';
 
 export interface IConfigState {
   url: string;
 }
 
-class Configuration extends React.Component<{}, IConfigState> {
-  constructor(props: {}) {
+class Configuration extends React.Component<any, IConfigState> {
+  constructor(props: any) {
     super(props);
     this.state = {
-      url: getBaseUrl() + "/messages?locale={locale}",
+      url: getBaseUrl() + '/messages?locale={locale}',
     };
   }
 
   public componentDidMount() {
-    microsoftTeams.initialize();
-    microsoftTeams.settings.registerOnSaveHandler((saveEvent) => {
-      microsoftTeams.settings.setSettings({
-        entityId: "Company_Communicator_App",
-        contentUrl: this.state.url,
-        suggestedDisplayName: "Company Communicator",
+    void app.initialize().then(() => {
+      pages.config.registerOnSaveHandler((saveEvent) => {
+        void pages.config.setConfig({
+          entityId: 'Company_Communicator_App',
+          contentUrl: this.state.url,
+          suggestedDisplayName: 'Company Communicator',
+        });
+        saveEvent.notifySuccess();
       });
-      saveEvent.notifySuccess();
-    });
 
-    microsoftTeams.settings.setValidityState(true);
+      pages.config.setValidityState(true);
+    });
   }
 
   public render(): JSX.Element {
     return (
-      <div className="configContainer">
+      <div className='configContainer'>
         <h3>Please click Save to get started.</h3>
       </div>
     );
