@@ -2,5 +2,6 @@
 // Licensed under the MIT License.
 
 export const getBaseUrl = (): string => {
-  return window.location.origin;
+  // return window.location.origin;
+  return 'https://c099-103-133-67-183.ngrok-free.app';
 };
