@@ -1,4 +1,4 @@
-import * as React from "react";
+import * as React from 'react';
 
 export const useInterval = (callback: any, delay: any) => {
   const savedCallback = React.useRef<any>();
@@ -15,12 +15,10 @@ export const useInterval = (callback: any, delay: any) => {
     };
 
     if (delay !== null) {
-      /* tslint:disable-next-line */
       const id = setInterval(tick, delay);
-      /* tslint:disable-next-line */
-      return () => clearInterval(id);
-    } else {
-      return;
+      return () => {
+        clearInterval(id);
+      };
     }
   }, [delay]);
 };
